feat: rename the `Mcu` component to `Mtb`, keep `Mcu` as a deprecated alias
