+++ .changeset/mcu-renamed-mtb.md
@@ -1,0 +1,5 @@
+---
+"material-theme-builder": minor
+---
+
+Rename the `Mcu` component to `Mtb`. `Mcu` still works as a deprecated alias of `Mtb` (same component, same props) and will be removed in the next major.
