+++ src/react.ts
@@ -4,5 +4,5 @@
 // callable from a server component. See tsup.config.ts.
 
 export { ExportButton } from "./ExportButton";
-export { Mcu } from "./Mcu";
 export { useMcu } from "./Mcu.context";
+export { Mcu, Mtb } from "./Mtb";
